Fixed tests and moved no pauschbetrag out of pauschbetrag
