+++ webapp/app/forms/flows/lotse_flow.py
@@ -361,17 +361,17 @@ class LotseMultiStepFlow(MultiStepFlow):
         """
         step_data = {}
         for attr in step.create_form(request.form, form_data).__dict__:
-            if attr in form_data or hasattr(step.form, attr):
-                field = getattr(step.form, attr)
-                # TODO: When the summary page is refactored we should merge _generate_value_representation & get_overview_value_representation
-                label, value = self._generate_value_representation(field, form_data.get(attr))
-                value = step.get_overview_value_representation(value)
-                if value is not None:
-                    step_data[label] = value
-            elif missing_fields and attr in missing_fields:
+            if missing_fields and attr in missing_fields:
                 field = getattr(step.form, attr)
                 label = field.kwargs['render_kw']['data_label']
                 step_data[label] = _l('form.lotse.missing_mandatory_field')
+            elif attr in form_data or hasattr(step.form, attr):
+                field = getattr(step.form, attr)
+                # TODO: When the summary page is refactored we should merge _generate_value_representation & get_overview_value_representation
+                label, value = self._generate_value_representation(field, form_data.get(attr))
+                value = step.get_overview_value_representation(value)                
+                if value is not None:
+                    step_data[label] = value
 
         return step_data
