+++ webapp/app/forms/steps/lotse/pauschbetrag.py
@@ -8,7 +8,6 @@ from wtforms.validators import InputRequired, ValidationError
 from wtforms import SelectField
 from flask_wtf.csrf import generate_csrf
 
-from app.forms.steps.lotse.fahrkostenpauschbetrag import calculate_fahrkostenpauschbetrag
 from app.model.components import PauschbetragProps
 from app.model.components.helpers import form_fields_dict
 from app.forms import SteuerlotseBaseForm
@@ -18,8 +17,6 @@ from app.forms.steps.lotse.personal_data import ShowPersonBPrecondition, StepFam
 from app.forms.steps.lotse.merkzeichen import StepMerkzeichenPersonA, StepMerkzeichenPersonB
 from app.forms.steps.lotse.utils import get_number_of_users
 from app.forms.steps.lotse.has_disability import HasDisabilityPersonAPrecondition, HasDisabilityPersonBPrecondition
-
-from app.model.components import NoPauschbetragProps
 
 
 def calculate_pauschbetrag(has_pflegegrad=None, disability_degree=None, has_merkzeichen_bl=False, has_merkzeichen_tbl=False, has_merkzeichen_h=False):
@@ -139,60 +136,6 @@ class HasPauschbetragClaimPersonBPrecondition(DisabilityModel):
         return values
 
 
-class HasNoPauschbetragOrFahrkostenpauschbetragClaimPersonAPrecondition(DisabilityModel):
-    _step_to_redirect_to = StepMerkzeichenPersonA.name
-    _message_to_flash = _l('form.lotse.skip_reason.has_pauschbetrag_claim')
-
-    @root_validator(skip_on_failure=True)
-    def has_to_have_pauschbetrag_and_fahrkostenpauschbetrag_not_0(cls, values):
-        pauschbetrag_claim = calculate_pauschbetrag(
-            has_pflegegrad=values.get('person_a_has_pflegegrad', None),
-            disability_degree=values.get('person_a_disability_degree', None),
-            has_merkzeichen_bl=values.get('person_a_has_merkzeichen_bl', False),
-            has_merkzeichen_tbl=values.get('person_a_has_merkzeichen_tbl', False),
-            has_merkzeichen_h=values.get('person_a_has_merkzeichen_h', False)
-        )
-        fahrkostenpauschbetrag_claim = calculate_fahrkostenpauschbetrag(
-            has_pflegegrad=values.get('person_a_has_pflegegrad', None),
-            disability_degree=values.get('person_a_disability_degree', None),
-            has_merkzeichen_bl=values.get('person_a_has_merkzeichen_bl', False),
-            has_merkzeichen_tbl=values.get('person_a_has_merkzeichen_bl', False),
-            has_merkzeichen_h=values.get('person_a_has_merkzeichen_bl', False),
-            has_merkzeichen_ag=values.get('person_a_has_merkzeichen_tbl', False),
-            has_merkzeichen_g=values.get('person_a_has_merkzeichen_h', False)
-        )
-
-        if pauschbetrag_claim != 0 or fahrkostenpauschbetrag_claim != 0:
-            raise ValidationError
-        return values
-
-
-class HasNoPauschbetragOrFahrkostenpauschbetragClaimPersonBPrecondition(DisabilityModel):
-    _step_to_redirect_to = StepMerkzeichenPersonB.name
-    _message_to_flash = _l('form.lotse.skip_reason.has_pauschbetrag_claim')
-
-    @root_validator(skip_on_failure=True)
-    def has_to_have_pauschbetrag_and_fahrkostenpauschbetrag_not_0(cls, values):
-        pauschbetrag_claim = calculate_pauschbetrag(
-            has_pflegegrad=values.get('person_b_has_pflegegrad', None),
-            disability_degree=values.get('person_b_disability_degree', None),
-            has_merkzeichen_bl=values.get('person_b_has_merkzeichen_bl', False),
-            has_merkzeichen_tbl=values.get('person_b_has_merkzeichen_tbl', False),
-            has_merkzeichen_h=values.get('person_b_has_merkzeichen_h', False)
-        )
-        fahrkostenpauschbetrag_claim = calculate_fahrkostenpauschbetrag(
-            has_pflegegrad=values.get('person_b_has_pflegegrad', None),
-            disability_degree=values.get('person_b_disability_degree', None),
-            has_merkzeichen_bl=values.get('person_b_has_merkzeichen_bl', False),
-            has_merkzeichen_tbl=values.get('person_b_has_merkzeichen_tbl', False),
-            has_merkzeichen_h=values.get('person_b_has_merkzeichen_h', False)
-        )
-
-        if pauschbetrag_claim != 0 or fahrkostenpauschbetrag_claim != 0:
-            raise ValidationError
-        return values
-
-
 class HasMerkzeichenPersonAPrecondition(BaseModel):
     _step_to_redirect_to = StepMerkzeichenPersonA.name
     _message_to_flash = _l('form.lotse.skip_reason.has_no_merkzeichen')
@@ -217,6 +160,7 @@ class HasMerkzeichenPersonAPrecondition(BaseModel):
         if not v and not any([values.get(merkzeichen_key) for merkzeichen_key in merkzeichen_keys]):
             raise ValueError
         return v
+
 
 class StepPauschbetragPersonA(StepPauschbetrag):
     name = 'person_a_requests_pauschbetrag'
@@ -343,57 +287,3 @@ class StepPauschbetragPersonB(StepPauschbetrag):
             has_merkzeichen_h=self.stored_data.get('person_b_has_merkzeichen_h', False)
         )
 
-
-class StepNoPauschbetragPersonA(LotseFormSteuerlotseStep):
-    name = 'person_a_no_pauschbetrag'
-    title = _l('form.lotse.no_pauschbetrag.person_a.title')
-    header_title = _l('form.lotse.mandatory_data.header-title')
-    preconditions = [HasDisabilityPersonAPrecondition,
-                     HasNoPauschbetragOrFahrkostenpauschbetragClaimPersonAPrecondition]
-
-    def _pre_handle(self):
-        self._set_multiple_texts()
-        super()._pre_handle()
-
-    def _set_multiple_texts(self):
-        num_of_users = get_number_of_users(self.render_info.stored_data)
-        self.render_info.step_title = ngettext('form.lotse.no_pauschbetrag.person_a.title',
-                                               'form.lotse.no_pauschbetrag.person_a.title',
-                                               num=num_of_users)
-
-    def render(self):
-        props_dict = NoPauschbetragProps(
-            step_header={
-                'title': str(self.title),
-            },
-            prev_url=self.render_info.prev_url,
-            next_url=self.render_info.next_url,
-        ).camelized_dict()
-
-        return render_template('react_component.html',
-                               component='NoPauschbetragPage',
-                               props=props_dict,
-                               form=self.render_info.form,
-                               header_title=_('form.lotse.header-title'))
-
-
-class StepNoPauschbetragPersonB(LotseFormSteuerlotseStep):
-    name = 'person_b_no_pauschbetrag'
-    title = _l('form.lotse.no_pauschbetrag.person_b.title')
-    header_title = _l('form.lotse.mandatory_data.header-title')
-    preconditions = [ShowPersonBPrecondition, HasDisabilityPersonBPrecondition, HasNoPauschbetragOrFahrkostenpauschbetragClaimPersonBPrecondition]
-
-    def render(self):
-        props_dict = NoPauschbetragProps(
-            step_header={
-                'title': str(self.title),
-            },
-            prev_url=self.render_info.prev_url,
-            next_url=self.render_info.next_url,
-        ).camelized_dict()
-
-        return render_template('react_component.html',
-                               component='NoPauschbetragPage',
-                               props=props_dict,
-                               form=self.render_info.form,
-                               header_title=_('form.lotse.header-title'))
