+++ webapp/tests/forms/steps/lotse/test_pauschbetrag.py
@@ -4,8 +4,7 @@ import pytest
 from pydantic import ValidationError
 
 from app.forms.steps.lotse.pauschbetrag import calculate_pauschbetrag, HasPauschbetragClaimPersonAPrecondition, \
-    HasPauschbetragClaimPersonBPrecondition, HasNoPauschbetragOrFahrkostenpauschbetragClaimPersonAPrecondition, \
-    HasNoPauschbetragOrFahrkostenpauschbetragClaimPersonBPrecondition, HasMerkzeichenPersonAPrecondition, \
+    HasPauschbetragClaimPersonBPrecondition,  HasMerkzeichenPersonAPrecondition, \
     HasMerkzeichenPersonBPrecondition
 
 class TestCalculatePauschbetrag:
@@ -219,87 +218,6 @@ class TestHasPauschbetragClaimPersonBPrecondition:
     def test_if_calculate_pauschbetrag_returns_number_other_than_zero_then_raise_no_error(self):
         with patch('app.forms.steps.lotse.pauschbetrag.calculate_pauschbetrag', MagicMock(return_value=1)):
             HasPauschbetragClaimPersonBPrecondition.parse_obj({})
-
-
-class TestHasNoPauschbetragOrFahrkostenpauschbetragClaimPersonAPrecondition:
-    def test_if_calculate_pauschbetrag_and_fahrkostenpauschbetrag_return_zero_then_raise_no_error(self):
-        with patch('app.forms.steps.lotse.pauschbetrag.calculate_pauschbetrag', MagicMock(return_value=0)), \
-                patch('app.forms.steps.lotse.pauschbetrag.calculate_fahrkostenpauschbetrag',
-                      MagicMock(return_value=0)):
-            HasNoPauschbetragOrFahrkostenpauschbetragClaimPersonAPrecondition.parse_obj({})
-
-    def test_if_calculate_pauschbetrag_and_fahrkostenpauschbetrag_return_number_other_than_zero_then_raise_validation_error(self):
-        with patch('app.forms.steps.lotse.pauschbetrag.calculate_pauschbetrag', MagicMock(return_value=1)), \
-                patch('app.forms.steps.lotse.pauschbetrag.calculate_fahrkostenpauschbetrag',
-                    MagicMock(return_value=1)):
-            with pytest.raises(ValidationError):
-                HasNoPauschbetragOrFahrkostenpauschbetragClaimPersonAPrecondition.parse_obj({})
-
-    def test_if_calculate_pauschbetrag_return_zero_and_fahrkostenpauschbetrag_return_number_other_than_raise_validation_error(self):
-        with patch('app.forms.steps.lotse.pauschbetrag.calculate_pauschbetrag', MagicMock(return_value=0)), \
-                patch('app.forms.steps.lotse.pauschbetrag.calculate_fahrkostenpauschbetrag',
-                    MagicMock(return_value=1)):
-            with pytest.raises(ValidationError):
-                HasNoPauschbetragOrFahrkostenpauschbetragClaimPersonAPrecondition.parse_obj({})
-
-    def test_if_calculate_pauschbetrag_return_number_other_than_zero_and_fahrkostenpauschbetrag_return_zero_then_raise_validation_error(self):
-        with patch('app.forms.steps.lotse.pauschbetrag.calculate_pauschbetrag', MagicMock(return_value=1)), \
-                patch('app.forms.steps.lotse.pauschbetrag.calculate_fahrkostenpauschbetrag',
-                    MagicMock(return_value=0)):
-            with pytest.raises(ValidationError):
-                HasNoPauschbetragOrFahrkostenpauschbetragClaimPersonAPrecondition.parse_obj({})
-
-    def test_if_calculate_pauschbetrag_returns_number_other_than_zero_then_raise_validation_error(self):
-        with patch('app.forms.steps.lotse.pauschbetrag.calculate_pauschbetrag',
-                   MagicMock(return_value=1)):
-            with pytest.raises(ValidationError):
-                HasNoPauschbetragOrFahrkostenpauschbetragClaimPersonAPrecondition.parse_obj({})
-
-    def test_if_calculate_fahrkostenpauschbetrag_returns_number_other_than_zero_then_raise_validation_error(self):
-        with patch('app.forms.steps.lotse.pauschbetrag.calculate_fahrkostenpauschbetrag',
-                   MagicMock(return_value=1)):
-            with pytest.raises(ValidationError):
-                HasNoPauschbetragOrFahrkostenpauschbetragClaimPersonAPrecondition.parse_obj({})
-
-
-class TestHasNoPauschbetragOrFahrkostenpauschbetragClaimPersonBPrecondition:
-    def test_if_calculate_pauschbetrag_and_fahrkostenpauschbetrag_return_zero_then_raise_no_error(self):
-        with patch('app.forms.steps.lotse.pauschbetrag.calculate_pauschbetrag', MagicMock(return_value=0)), \
-                patch('app.forms.steps.lotse.pauschbetrag.calculate_fahrkostenpauschbetrag',
-                      MagicMock(return_value=0)):
-            HasNoPauschbetragOrFahrkostenpauschbetragClaimPersonBPrecondition.parse_obj({})
-
-    def test_if_calculate_pauschbetrag_and_fahrkostenpauschbetrag_return_number_other_than_zero_then_raise_validation_error(self):
-        with patch('app.forms.steps.lotse.pauschbetrag.calculate_pauschbetrag', MagicMock(return_value=1)), \
-        patch('app.forms.steps.lotse.pauschbetrag.calculate_fahrkostenpauschbetrag',
-                MagicMock(return_value=1)):
-            with pytest.raises(ValidationError):
-                HasNoPauschbetragOrFahrkostenpauschbetragClaimPersonBPrecondition.parse_obj({})
-
-    def test_if_calculate_pauschbetrag_return_zero_and_fahrkostenpauschbetrag_return_number_other_than_raise_validation_error(self):
-        with patch('app.forms.steps.lotse.pauschbetrag.calculate_pauschbetrag', MagicMock(return_value=0)):
-            with patch('app.forms.steps.lotse.pauschbetrag.calculate_fahrkostenpauschbetrag',
-                    MagicMock(return_value=1)):
-                with pytest.raises(ValidationError):
-                    HasNoPauschbetragOrFahrkostenpauschbetragClaimPersonBPrecondition.parse_obj({})
-
-    def test_if_calculate_pauschbetrag_return_number_other_than_zero_and_fahrkostenpauschbetrag_return_zero_then_raise_validation_error(self):
-        with patch('app.forms.steps.lotse.pauschbetrag.calculate_pauschbetrag', MagicMock(return_value=1)):
-            with patch('app.forms.steps.lotse.pauschbetrag.calculate_fahrkostenpauschbetrag',
-                    MagicMock(return_value=0)):
-                with pytest.raises(ValidationError):
-                    HasNoPauschbetragOrFahrkostenpauschbetragClaimPersonBPrecondition.parse_obj({})
-
-    def test_if_calculate_pauschbetrag_returns_number_other_than_zero_then_raise_validation_error(self):
-        with patch('app.forms.steps.lotse.pauschbetrag.calculate_pauschbetrag', MagicMock(return_value=1)):
-            with pytest.raises(ValidationError):
-                HasNoPauschbetragOrFahrkostenpauschbetragClaimPersonBPrecondition.parse_obj({})
-
-    def test_if_calculate_fahrkostenpauschbetrag_returns_number_other_than_zero_then_raise_validation_error(self):
-        with patch('app.forms.steps.lotse.pauschbetrag.calculate_fahrkostenpauschbetrag',
-                   MagicMock(return_value=1)):
-            with pytest.raises(ValidationError):
-                HasNoPauschbetragOrFahrkostenpauschbetragClaimPersonBPrecondition.parse_obj({})
 
 class TestHasMerkzeichenPersonAPrecondition:
